requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
boto3==1.35.29
pyyaml==6.0.1
aiohttp==3.10.5
click==8.1.7
backoff==2.2.1
